Streams suites: absorb streams-test.js and streams-backpressure-test.js strays
