+++ src/tests/streams/readable/readable-cpp-pedantic.wd-test
@@ -21,6 +21,7 @@ const unitTests :Workerd.Config = (
           "set_tostring_tag",
           "writable_stream_spec_compliant_writer",
           "pedantic_wpt",
+          "enable_weak_ref",
         ],
         bindings = [
           ( name = "SELF", service = "streams-readable-cpp-pedantic" ),
