--- src/workerd/api/tests/streams-backpressure-test.wd-test
@@ -1,14 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "streams-backpressure-test",
-      worker = (
-        modules = [
-          (name = "worker", esModule = embed "streams-backpressure-test.js")
-        ],
-        compatibilityFlags = ["nodejs_compat", "streams_enable_constructors", "transformstream_enable_standard_constructor"],
-      )
-    ),
-  ],
-);
