--- src/workerd/api/tests/streams-test.wd-test
@@ -1,26 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  v8Flags = ["--expose-gc"],
-  services = [
-    ( name = "streams-test",
-      worker = (
-        modules = [
-          (name = "worker", esModule = embed "streams-test.js")
-        ],
-        compatibilityFlags = [
-          "nodejs_compat",
-          "internal_stream_byob_return_view",
-          "streams_enable_constructors",
-          "transformstream_enable_standard_constructor",
-          "streams_byob_reader_detaches_buffer",
-          "enhanced_error_serialization",
-          "enable_weak_ref",
-        ],
-        bindings = [
-          (name = "subrequest", service = "streams-test")
-        ]
-      )
-    ),
-  ],
-);
